gate-bf16 (flag) + drop redundant attn_sink fp32 copy + lazy gen_nsa_ks_ke
    16k 352.8->368.1 (+4.3%), 64k 86.4->90.0 (+4.2%). needle@16.5k PASS; gsm8k 8-shot 300q acc=0.900/0-invalid (== fp32 baseline ~0.
902) with LIGHTLLM_DSV4_GATE_BF16=1.

+++ lightllm/models/deepseek_v4/layer_weights/transformer_layer_weight.py
@@ -189,14 +189,13 @@ class DeepseekV4TransformerLayerWeight(TransformerLayerWeight):
     # ------------------------------------------------------------------ moe
     def _init_moe(self):
         p = f"{self.prefix}.ffn"
-        # router gate (replicated). Stored as fp32: the topk_hash_softplus_sqrt router wants fp32 logits,
-        # so keep the gate matmul in fp32 — but store the (constant) weight as fp32 once here instead of
-        # re-casting it to fp32 on every forward in _ffn.
+        # Router gate in bf16 (matches the sglang/vLLM DeepSeek references, which run the gate GEMM in
+        # the model dtype); the bf16 GEMM output is cast back to fp32 in _ffn for topk_hash_softplus_sqrt.
         self.gate_weight_ = ROWMMWeight(
             in_dim=self.hidden,
             out_dims=[self.n_routed_experts],
             weight_names=f"{p}.gate.weight",
-            data_type=torch.float32,
+            data_type=torch.bfloat16,
             quant_method=None,
             tp_rank=0,
             tp_world_size=1,
